fix legacy errors - REMOVAL of LAZY.nvim
